Created AdjacentGetter class to merge two identical methods in OpponentSetInteractor and ChestInteractor.
