+++ src/main/java/com/mg105/use_cases/OpponentSetInteractor.java
@@ -14,27 +14,6 @@ public class OpponentSetInteractor {
     }
 
     /**
-     * Creates a list of four adjacent coordinates on the map next to the WalkingCharacter.
-     * @return the list of adjacent coordinates.
-     */
-    //This method is essentially the same as the one with the same name in ChestInteractor.
-    //Both methods could be replaced onto an external class.
-    public static List<Point> getAdjacents(Point coord) {
-        int xCoord = coord.x;
-        int yCoord = coord.y;
-        ArrayList<Point> adjacents = new ArrayList<>();
-        Point adjOne = new Point((xCoord - 1), yCoord);
-        Point adjTwo = new Point((xCoord + 1), yCoord);
-        Point adjThree = new Point(xCoord, (yCoord - 1));
-        Point adjFour = new Point(xCoord, (yCoord + 1));
-        adjacents.add(adjOne);
-        adjacents.add(adjTwo);
-        adjacents.add(adjThree);
-        adjacents.add(adjFour);
-        return adjacents;
-    }
-
-    /**
      * Checks if a given spot on the map contains an OpponentSet.
      * @return the OpponentSet if opponent is found, or null if it is not.
      */
@@ -51,12 +30,11 @@ public class OpponentSetInteractor {
      * Checks if an OpponentSet is adjacent to the player character.
      * @return the OpponentSet if the opponent is found, or null if it is not.
      */
-
     public OpponentSet verifyOpponent() {
 
         Point charPosition = state.getWalkingCharacter().getCharPosition();
         List<OpponentSet> opponents = state.getCurrentRoom().getOpponents();
-        List<Point> adjacents = getAdjacents(charPosition);
+        List<Point> adjacents = AdjacentGetter.getAdjacents(charPosition);
 
         for (Point adjacentSpot : adjacents) {
             OpponentSet result = compareLists(adjacentSpot, opponents);
