+++ src/test/java/com/mg105/entities/ChestTest.java
@@ -2,6 +2,7 @@ package com.mg105.entities;
 
 import com.mg105.entities.items.HealthPotion;
 import com.mg105.outputds.ItemDetails;
+import com.mg105.use_cases.ChestInteractor;
 import com.mg105.use_cases.Inventory.InventoryInteractor;
 import com.mg105.use_cases.Inventory.InventoryPresenterInterface;
 import com.mg105.utils.PartyConstants;
